Address review comments on bulk classification routes
- Resolve grid filters through grid_filter_sample_ids.build_sample_ids_query
  so annotation-grid embedding regions are honoured.
- Validate the parent collection via get_and_validate_collection_id (404
  instead of 400 for unknown collections).
- Use db_array.in_array for the explicit sample ID selection.
- Rename the routes to .../classifications/bulk_create[_by_filter].
- Move the tests into test_create_annotation.py to mirror the source file.
