+++ lightly_studio/src/lightly_studio/api/routes/api/annotations/create_annotation.py
@@ -2,29 +2,32 @@
 
 from __future__ import annotations
 
+from collections.abc import Sequence
 from typing import Annotated
 from uuid import UUID
 
-from fastapi import APIRouter, Path
+from fastapi import APIRouter, Depends, Path
 from fastapi.params import Body
 from pydantic import BaseModel
-from sqlmodel import Session, col, select
+from sqlmodel import col, select
 from sqlmodel.sql.expression import SelectOfScalar
 
+from lightly_studio.api.routes.api.collection import get_and_validate_collection_id
 from lightly_studio.api.routes.api.status import HTTP_STATUS_CREATED
+from lightly_studio.database import db_array
 from lightly_studio.database.db_manager import SessionDep
 from lightly_studio.models.annotation.annotation_base import (
     AnnotationBaseTable,
     AnnotationType,
     AnnotationView,
 )
+from lightly_studio.models.collection import CollectionTable
 from lightly_studio.models.sample import SampleTable
-from lightly_studio.resolvers import annotation_resolver, embedding_region_resolver
+from lightly_studio.resolvers import annotation_resolver, grid_filter_sample_ids
 from lightly_studio.resolvers.annotation_resolver.bulk_create_classifications import (
     BulkCreateClassificationsResult,
 )
 from lightly_studio.resolvers.grid_filter import GridFilter
-from lightly_studio.resolvers.image_filter import ImageFilter
 from lightly_studio.services import annotations_service
 from lightly_studio.services.annotations_service.create_annotation import AnnotationCreateParams
 
@@ -87,13 +90,15 @@ def create_annotation(
 
 
 @create_annotation_router.post(
-    "/annotations/classifications/bulk",
+    "/annotations/classifications/bulk_create",
     response_model=BulkCreateClassificationsResult,
     status_code=HTTP_STATUS_CREATED,
 )
 def bulk_create_classifications(
-    collection_id: Annotated[
-        UUID, Path(title="collection Id", description="The ID of the collection")
+    collection: Annotated[
+        CollectionTable,
+        Path(title="collection Id"),
+        Depends(get_and_validate_collection_id),
     ],
     session: SessionDep,
     body: Annotated[BulkCreateClassificationsInput, Body()],
@@ -101,9 +106,9 @@ def bulk_create_classifications(
     """Bulk-create classification annotations for selected samples."""
     return annotation_resolver.bulk_create_classifications(
         session=session,
-        parent_collection_id=collection_id,
+        parent_collection_id=collection.collection_id,
         parent_sample_ids_query=_build_sample_ids_query(
-            collection_id=collection_id,
+            collection_id=collection.collection_id,
             sample_ids=body.sample_ids,
         ),
         class_name=body.class_name,
@@ -112,27 +117,28 @@ def bulk_create_classifications(
 
 
 @create_annotation_router.post(
-    "/annotations/classifications/bulk_by_filter",
+    "/annotations/classifications/bulk_create_by_filter",
     response_model=BulkCreateClassificationsResult,
     status_code=HTTP_STATUS_CREATED,
 )
 def bulk_create_classifications_by_filter(
-    collection_id: Annotated[
-        UUID, Path(title="collection Id", description="The ID of the collection")
+    collection: Annotated[
+        CollectionTable,
+        Path(title="collection Id"),
+        Depends(get_and_validate_collection_id),
     ],
     session: SessionDep,
     body: Annotated[BulkCreateClassificationsByFilterInput, Body()],
 ) -> BulkCreateClassificationsResult:
     """Bulk-create classification annotations for filtered samples."""
-    grid_filter = _resolve_image_filter_embedding_region(
-        session=session,
-        collection_id=collection_id,
-        grid_filter=body.filter,
-    )
     return annotation_resolver.bulk_create_classifications(
         session=session,
-        parent_collection_id=collection_id,
-        parent_sample_ids_query=grid_filter.build_sample_ids_query(collection_id=collection_id),
+        parent_collection_id=collection.collection_id,
+        parent_sample_ids_query=grid_filter_sample_ids.build_sample_ids_query(
+            session=session,
+            collection_id=collection.collection_id,
+            grid_filter=body.filter,
+        ),
         class_name=body.class_name,
         annotation_collection_name=body.annotation_collection_name,
     )
@@ -140,30 +146,10 @@ def bulk_create_classifications_by_filter(
 
 def _build_sample_ids_query(
     collection_id: UUID,
-    sample_ids: list[UUID],
+    sample_ids: Sequence[UUID],
 ) -> SelectOfScalar[UUID]:
     return (
         select(SampleTable.sample_id)
         .where(SampleTable.collection_id == collection_id)
-        .where(col(SampleTable.sample_id).in_(sample_ids))
+        .where(db_array.in_array(column=col(SampleTable.sample_id), values=sample_ids))
     )
-
-
-def _resolve_image_filter_embedding_region(
-    session: Session,
-    collection_id: UUID,
-    grid_filter: GridFilter,
-) -> GridFilter:
-    if (
-        isinstance(grid_filter, ImageFilter)
-        and grid_filter.sample_filter is not None
-        and grid_filter.sample_filter.embedding_region is not None
-    ):
-        grid_filter.sample_filter.region_sample_ids = (
-            embedding_region_resolver.get_sample_ids_in_region(
-                session=session,
-                collection_id=collection_id,
-                region=grid_filter.sample_filter.embedding_region,
-            )
-        )
-    return grid_filter
